requirements.txt
PyYAML==5.3.1
sanic==20.3.0